define popup
I added a page called /info/mediaDocumentation.cfm to popup from the
Define link on both the edit block of Media.cfm and create block of
Media.cfm page. I added a function to the top of Media.cfm to get the
popup to work. I should create a new style sheet for the new styles and
consolidate all the "old" styles that are needed--for now they are on
the popup page.