ISSUE: 264  PURPOSE:  Needed to tweak code and table values for biol relationships
DESCRIPTION:  The values that were equal on both sides were not showing
because we had a line in the code to that effect.  Also, I added a value
of "hide" in the inverse code table but that does not hide the initial
value so I still had to add <> 'cloned from record' in the block
starting at line 251.  This should be resolved later so that there is a
fix for both sides without having to single out individual values.  Code
table still needs to be done.