ISSUE:  Added padding to .detailCell class--GUIDs
Added padding of 7px to the bottom of the text of the class .detailCell used for the GUID element on includes/style.css.  This class is also used on the publication page and possibly in other contexts so I didn't want to change it too much.
The margin doesn't work when the class is used in a <td> but <br>'s could be put in between the <tr>'s to mimic a margin.